+++ walkability/components/wellness/wellness_artifacts.py
@@ -11,8 +11,8 @@ from pydantic_extra_types.color import Color
 
 from walkability.components.utils.geometry import get_buffered_aoi
 from walkability.components.utils.misc import generate_colors
-from walkability.components.utils.ors_settings import ORSSettings
-from walkability.components.wellness.benches_and_drinking_water import PointsOfInterest, get_pois
+from walkability.components.wellness.benches_and_drinking_water import PointsOfInterest, distance_enrich_paths
+from walkability.core.settings import ORSSettings
 
 log = logging.getLogger(__name__)
 N_BINS = 5
@@ -27,18 +27,18 @@ def compute_wellness_artifacts(
     resources: ComputationResources,
 ) -> list[_Artifact]:
     artifacts = []
-    for poi in [PointsOfInterest.DRINKING_WATER, PointsOfInterest.SEATING]:
-        log.debug(f'Computing Wellness for {poi}')
-        max_walking_distance = max_walking_distance_map[poi]
-        step_size = int(max_walking_distance / N_BINS)
-        bins = [x for x in range(step_size, int(max_walking_distance) + 1, step_size)]
+    for poi_type in [PointsOfInterest.DRINKING_WATER, PointsOfInterest.SEATING]:
+        log.debug(f'Computing Wellness for {poi_type}')
+        max_walking_distance = max_walking_distance_map[poi_type]
+        bin_size = int(max_walking_distance / N_BINS)
+        bins = [x for x in range(bin_size, int(max_walking_distance) + 1, bin_size)]
         max_walking_distance = max(bins)
 
         buffered_aoi = get_buffered_aoi(aoi, max_walking_distance)
-        enriched_paths = get_pois(
+        enriched_paths = distance_enrich_paths(
             paths=paths,
             aoi=buffered_aoi,
-            poi=poi,
+            poi_type=poi_type,
             bins=bins,
             ohsome_client=ohsome_client,
             ors_settings=ors_settings,
@@ -46,7 +46,12 @@ def compute_wellness_artifacts(
         enriched_paths = enriched_paths.clip(aoi)
 
         isodistance_artifact = build_isodistance_artifact(
-            resources=resources, data=enriched_paths, max_walking_distance=max_walking_distance, poi=poi, bins=bins
+            resources=resources,
+            data=enriched_paths,
+            max_walking_distance=max_walking_distance,
+            poi_type=poi_type,
+            bins=bins,
+            max_isochrone_request=ors_settings.ors_isochrone_max_request_number,
         )
         artifacts.append(isodistance_artifact)
 
@@ -56,12 +61,13 @@ def compute_wellness_artifacts(
 def build_isodistance_artifact(
     resources: ComputationResources,
     data: gpd.GeoDataFrame,
-    poi: PointsOfInterest,
+    poi_type: PointsOfInterest,
     bins: list[int],
     max_walking_distance: float,
+    max_isochrone_request: int,
 ) -> _Artifact:
     log.debug('Building isodistance artifact')
-    cleaned_data = clean_data(data, max_walking_distance, min_value=min(bins), poi=poi)
+    cleaned_data = clean_data(data, max_walking_distance, min_value=min(bins), poi_type=poi_type)
 
     legend = {}
     unique_labels = cleaned_data.sort_values('value').label.unique()
@@ -71,33 +77,37 @@ def build_isodistance_artifact(
 
     return create_geojson_artifact(
         features=data.geometry,
-        layer_name=f'Distance to {poi.value.title()}',
-        caption=f'How far is it to {poi.value.capitalize()}?',
+        layer_name=f'Distance to {poi_type.value.title()}',
+        caption=f'How far is it to {poi_type.value.capitalize()}?',
+        description=f'Note that the distance to {poi_type.value.title()} will be estimated based on euclidean distance if '
+        f'more than {max_isochrone_request} {poi_type.value.title()} exist in the area of interest. Otherwise '
+        f'more precise isochrones will be used, that reflect the real walking distance. Therefore, to '
+        f'ensure isochrone usage, make sure to use a consice area of interest.',
         color=cleaned_data['color'].to_list(),
         label=cleaned_data['label'].to_list(),
         legend_data=legend,
         resources=resources,
-        filename=f'isodistance_{poi.value.replace(" ", "_")}',
+        filename=f'isodistance_{poi_type.value.replace(" ", "_")}',
     )
 
 
 def clean_data(
-    data: gpd.GeoDataFrame, max_walking_distance: float, min_value: float, poi: PointsOfInterest
+    data: gpd.GeoDataFrame, max_walking_distance: float, min_value: float, poi_type: PointsOfInterest
 ) -> gpd.GeoDataFrame:
-    data['label'] = data.apply(assign_label, poi=poi, max_walking_distance=max_walking_distance, axis=1)
+    data['label'] = data.apply(assign_label, poi_type=poi_type, max_walking_distance=max_walking_distance, axis=1)
 
-    data = assign_color(data, max_walking_distance=max_walking_distance, min_value=min_value, poi=poi)
+    data = assign_color(data, max_walking_distance=max_walking_distance, min_value=min_value, poi_type=poi_type)
 
     return data
 
 
 def assign_color(
-    data: gpd.GeoDataFrame, max_walking_distance: float, min_value: float, poi: PointsOfInterest
+    data: gpd.GeoDataFrame, max_walking_distance: float, min_value: float, poi_type: PointsOfInterest
 ) -> gpd.GeoDataFrame:
     data['color'] = generate_colors(
         data.value, 'coolwarm', min_value, max_value=max_walking_distance, bad_color=Color('red').as_hex()
     )
-    match poi:
+    match poi_type:
         case PointsOfInterest.SEATING:
             point_color = Color('brown')
         case PointsOfInterest.DRINKING_WATER:
@@ -109,9 +119,9 @@ def assign_color(
     return data
 
 
-def assign_label(row: pd.Series, poi: PointsOfInterest, max_walking_distance: float) -> str:
+def assign_label(row: pd.Series, poi_type: PointsOfInterest, max_walking_distance: float) -> str:
     match row.geometry.geom_type:
         case 'Point':
-            return poi.value
+            return poi_type.value
         case _:
             return f'> {int(max_walking_distance)}m' if np.isnan(row['value']) else f'< {int(row["value"])}m'